fix: mock ref

+++ mobile/src/screens/App/SettingsScreen.tsx
@@ -5,6 +5,7 @@ import { Divider, Avatar, Button } from 'react-native-paper';
 
 import { IResponse, IMessage, IReference, IContact, IDocumentType, IGood, IRemain } from '../../../../common';
 import SettingsItem from '../../components/SettingsItem';
+import config from '../../config';
 import { timeout } from '../../helpers/utils';
 import { useAuthStore, useAppStore, useServiceStore } from '../../store';
 
@@ -65,33 +66,39 @@ const SettingsScreen = () => {
     }
 
     const getMessages = async () => {
-      try {
-        const response = await timeout<IResponse<IMessage[]>>(5000, apiService.data.getMessages(companyID));
-        if (!response.result) {
-          Alert.alert('Запрос не был отправлен', '', [{ text: 'Закрыть', onPress: () => ({}) }]);
-          return;
-        }
-        const messages = response.data.filter((message) => message.body.type === 'data');
-        if (messages.length !== 0) {
-          const sortMessages = messages.sort((curr, next) => next.head.dateTime.localeCompare(curr.head.dateTime));
-          /* Обработка данных по справочникам */
-          const ref = (sortMessages[0].body.payload as unknown) as IReference[];
-          const contacts = ref.find((itm) => itm.name === 'contacts')?.data as IContact[];
-          appActions.setContacts(contacts);
-          const documentTypes = ref.find((itm) => itm.name === 'documenttypes')?.data as IDocumentType[];
-          appActions.setDocumentTypes(documentTypes);
-          const goods = ref.find((itm) => itm.name === 'goods')?.data as IGood[];
-          appActions.setGoods(goods);
-          const remains = (ref.find((itm) => itm.name === 'remains')?.data as unknown) as IRemain[];
-          appActions.setRemains(remains);
+      if (config.debug.useMockup) {
+        const mockRef = (await import('../../mockData/References.json')) as IReference[];
+        const mockContacts = mockRef.find((itm) => itm.name === 'contacts')?.data as IContact[];
+        appActions.setContacts(mockContacts);
+      } else {
+        try {
+          const response = await timeout<IResponse<IMessage[]>>(5000, apiService.data.getMessages(companyID));
+          if (!response.result) {
+            Alert.alert('Запрос не был отправлен', '', [{ text: 'Закрыть', onPress: () => ({}) }]);
+            return;
+          }
+          const messages = response.data.filter((message) => message.body.type === 'data');
+          if (messages.length !== 0) {
+            const sortMessages = messages.sort((curr, next) => next.head.dateTime.localeCompare(curr.head.dateTime));
+            /* Обработка данных по справочникам */
+            const ref = (sortMessages[0].body.payload as unknown) as IReference[];
+            const contacts = ref.find((itm) => itm.name === 'contacts')?.data as IContact[];
+            appActions.setContacts(contacts);
+            const documentTypes = ref.find((itm) => itm.name === 'documenttypes')?.data as IDocumentType[];
+            appActions.setDocumentTypes(documentTypes);
+            const goods = ref.find((itm) => itm.name === 'goods')?.data as IGood[];
+            appActions.setGoods(goods);
+            const remains = (ref.find((itm) => itm.name === 'remains')?.data as unknown) as IRemain[];
+            appActions.setRemains(remains);
 
-          apiService.data.deleteMessage(companyID, messages[0].head.id);
-          // appActions.setReferences((sortMessages[0].body.payload as unknown) as IReference[]);
+            apiService.data.deleteMessage(companyID, messages[0].head.id);
+            // appActions.setReferences((sortMessages[0].body.payload as unknown) as IReference[]);
+          }
+        } catch (err) {
+          Alert.alert('Ошибка!', err.message, [{ text: 'Закрыть', onPress: () => ({}) }]);
         }
-        Alert.alert('Данные получены', 'Справочники обновлены', [{ text: 'Закрыть', onPress: () => ({}) }]);
-      } catch (err) {
-        Alert.alert('Ошибка!', err.message, [{ text: 'Закрыть', onPress: () => ({}) }]);
       }
+      Alert.alert('Данные получены', 'Справочники обновлены', [{ text: 'Закрыть', onPress: () => ({}) }]);
     };
 
     getMessages();
